Let one bad tag cost one tag, keep numberless volumes, decode fractional counts
W14: tags_v2 was decoded as one array under try?. A single tag with a null
name emptied all 146 of Solo Leveling's, and the flat fallback list hid that
the rich ones had gone. Decoded element by element now, through a lenient
container helper, so a malformed tag costs itself.
  (series.richTags.map(\.name) → []) == ["Action", "Murim"]

W5: volumes(from:) promised in its own comment that a box set or a side story
"keeps the API's own order at the end", then `continue`d past them, and the
nil-handling in the sort below was dead work. They now sit under "Other
editions" at the end. The test that asserted they were dropped was written to
the bug and is flipped.

W17: CommunityPulse decoded four figures as Int against a schema that types
all of them as number, beside a sibling measured to arrive fractional. Int
throws on 303800.5 and a throw hid the whole card. All six are Double, rounded
where shown.

W15 is a false positive: preferredCover returning nil is how the series' own
cover — MangaBaka's pick, usually the native volume — stands. The caller falls
back to it. A comment on the function records that so it is not re-proposed.

+++ MangaBaka/Core/Model/CommunityPulse.swift
@@ -11,10 +11,13 @@ import Foundation
 /// decoded: Abdi asked for the database's pulse, not a leaderboard, and a
 /// contribution count on a screen nobody contributes from is vanity.
 struct CommunityPulse: Codable, Equatable, Sendable {
-    let activeSeriesCount: Int
-    let activeSeriesCountPrevWeek: Int
-    let registeredUserCount: Int
-    let registeredUserCountPrevWeek: Int
+    /// Doubles, all six: the schema types every figure as `number`, and
+    /// `chapters_read_count` is measured to arrive fractional. `Int` throws
+    /// on 290.0, and a throw here hid the whole card. Rounded where shown.
+    let activeSeriesCount: Double
+    let activeSeriesCountPrevWeek: Double
+    let registeredUserCount: Double
+    let registeredUserCountPrevWeek: Double
     let chaptersReadCount: Double
     let chaptersReadCountPrevWeek: Double
 
@@ -43,14 +46,14 @@ struct CommunityPulse: Codable, Equatable, Sendable {
             ),
             figure(
                 id: "series",
-                now: Double(activeSeriesCount),
-                then: Double(activeSeriesCountPrevWeek),
+                now: activeSeriesCount,
+                then: activeSeriesCountPrevWeek,
                 label: "series in the database"
             ),
             figure(
                 id: "readers",
-                now: Double(registeredUserCount),
-                then: Double(registeredUserCountPrevWeek),
+                now: registeredUserCount,
+                then: registeredUserCountPrevWeek,
                 label: "people keeping libraries"
             )
         ]
